+++ modules/@apostrophecms/schema/ui/apos/components/AposInputArea.vue
@@ -5,6 +5,8 @@
     :uid="uid"
     :display-options="displayOptions"
     :modifiers="[...modifiers, 'full-width']"
+    :items="next.items"
+    :meta="areaMeta"
   >
     <template #body>
       <!-- data-apos-schema-area lets all the child areas know that this area is in a schema (which is in a modal)
@@ -21,6 +23,7 @@
           :id="next._id"
           :options="field.options"
           :items="next.items"
+          :meta="areaMeta"
           :choices="choices"
           :field-id="field._id"
           :field="field"